getService/requirements.txt
connexion
PyMySQL
google-cloud-error-reporting
google-cloud-logging
